% $Id: Calendars.Rd 64 2009-04-10 20:24:33Z edd $
\name{GUI functions}
\alias{discountBuilder}
\alias{fittedBondBuilder}
\alias{bondGUI}
\title{GUI for RQuantLib}
\description{
These functions provide a friendly interface for fixed income functions
in RQuantLib.
}
\usage{
discountBuilder()
fittedBondBuilder()
bondGUI()
}
\references{\url{http://quantlib.org} for details on \code{QuantLib}.}
\author{Dirk Eddelbuettel \email{edd@debian.org} for the \R interface;
  Khanh Nguyen \email{knguyen@cs.umb.edu} for the implementation
}
\note{The interface might change in future release as \code{QuantLib}
  stabilises its own API.}
